🔨modify:  파일 컴포넌트 분할 - 모집 인원 컴포넌트로 분할 - 진행 방식 컴포넌트로 분할 - 모집 인원 컴포넌트로 분할
